+++ src/routes/MarketDetails/store/actions/buyShares.js
@@ -3,13 +3,13 @@ import uuid from 'uuid/v4'
 import * as api from 'api'
 import { fetchMarket, buyShares } from 'routes/MarketDetails/api'
 import Decimal from 'decimal.js'
-import { startLog, closeLog, closeEntrySuccess, closeEntryError } from 'store/actions//transactions'
+import { startLog, closeLog, closeEntrySuccess, closeEntryError } from 'routes/Transactions/store/actions//transactions'
 import { openModal, closeModal } from 'store/actions//modal'
 import gaSend from 'utils/analytics/gaSend'
-import { receiveEntities, updateEntity } from 'store/actions//entities'
+import { receiveEntities, updateEntity } from 'store/actions/entities'
 import { MAX_ALLOWANCE_WEI, TRANSACTION_COMPLETE_STATUS } from 'utils/constants'
 import { SETTING_ALLOWANCE, DEPOSIT } from 'utils/transactionExplanations'
-import { TRANSACTION_EVENTS_GENERIC, TRANSACTION_STAGES } from 'store/actions//market'
+import { TRANSACTION_EVENTS_GENERIC, TRANSACTION_STAGES } from 'store/actions/market'
 
 /**
  * Buy shares on specific market
