+++ src/routes/MarketDetails/store/actions/sellShares.js
@@ -2,13 +2,13 @@ import allowedRangePrice from 'utils/marginPrice'
 import uuid from 'uuid/v4'
 import * as api from 'api'
 import Decimal from 'decimal.js'
-import { startLog, closeLog, closeEntrySuccess, closeEntryError } from 'store/actions//transactions'
-import { openModal, closeModal } from 'store/actions//modal'
+import { startLog, closeLog, closeEntrySuccess, closeEntryError } from 'routes/Transactions/store/actions/transactions'
+import { openModal, closeModal } from 'store/actions/modal'
 import gaSend from 'utils/analytics/gaSend'
 import { receiveEntities, updateEntity } from 'store/actions//entities'
 import { MAX_ALLOWANCE_WEI, TRANSACTION_COMPLETE_STATUS } from 'utils/constants'
 import { SETTING_ALLOWANCE, SELL } from 'utils/transactionExplanations'
-import { TRANSACTION_EVENTS_GENERIC, TRANSACTION_STAGES } from 'store/actions//market'
+import { TRANSACTION_EVENTS_GENERIC, TRANSACTION_STAGES } from 'store/actions/market'
 import { sellShares, fetchMarket } from '../../api'
 /**
  * Sell shares on a specific market
