+++ src/store/reducers/index.js
@@ -3,11 +3,11 @@ import { routerReducer } from 'react-router-redux'
 import { reducer as formReducer } from 'redux-form'
 import integrations from 'integrations/store/reducers'
 import users from 'routes/Scoreboard/store/reducers/users'
+import transactions from 'routes/Transactions/store/reducers/transactions'
 import market, { REDUCER_ID } from 'store/reducers/market'
 import { isFeatureEnabled } from 'utils/features'
 import entities from './entities'
 import modal from './modal'
-import transactions from './transactions'
 import blockchain from './blockchain'
 import notifications from './notifications'
